fix: commentの作成、取得をstore経由でおこなうように
